+++ R/helpfunctions.R
@@ -209,7 +209,7 @@ create.frq.df <- function(x,
   # -------------------------------------
   if (!is.infinite(wrap.labels) && !is.null(labels)) {
     if (anyNA(labels)) labels <- stats::na.omit(labels)
-    labels <- sjmisc::word_wrap(labels, wrap.labels)
+    # labels <- sjmisc::word_wrap(labels, wrap.labels)
   }
   # -------------------------------------
   # return results