+++ NEWS.md
@@ -13,6 +13,7 @@
 * Minor fix in `tab_model()` for mixed models that can't compute R2.
 * The `plot_likert()` option `reverse.scale = TRUE` resulted in `values = "sum.inside"` being outside and the other way around. This is fixed now. 
 * `view_df()` mixed up labels and frequency values when value labels were present, but no such values were in the data.
+* Argument `wrap.labels` in `plot_frq()` did not properly work for factor levels.
 
 # sjPlot 2.6.3